requirement.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
cassandra-driver==3.25.0
click==8.1.2
Django==4.0.3
django-cassandra-engine==1.7.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
geomet==0.2.1.post1
pycodestyle==2.8.0
pytz==2022.1
scylla-driver==3.25.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
